Solve Timeline
Major, major milestone achieved!!! This Timeline can literally solve any equation for any variable.

Just kidding, the equation has to be linear in the target variable, and the variable can only appear once as it cannot solve quadratics or even combine linear terms. But still, I am very proud of this capability. It can solve for it no matter where it is, and will automatically reduce down any arithmetic as it appears.