Working on Functions and glyph getting
The whole way I've done glyph getting is crazy and nonrecursive. There's an argument that I have to do it this way to avoid the parentheses problems, but I think I will be avoiding those problems altogether with a different latex package. So I'm gonna redo it recursively where each expression is only responsible for handling one address character at a time. But it's not done yet