addigital dreamm modification
putting in method I was thinking about while away from home. Clearly Expressions should get_glyphs in a recursive manner, and ignore the edge cases where latex changes the number of submobjects in parentheses and such depending on how the expression is situated inside another. I will deal with this later in latex and not in python.