Auto Parentheses Animating
Primitive but successful implementation of get_addressmap wrapper automatically adding entries for fading in and out parentheses so that they need not be explicitly mentioned in the naive addressmap. Currently it just fades all parentheses it can find in and out. A smarter version will match pre and post addresses and check if the parentheses have changed and leave it alone if not.