Match Expresion Upgrade
Added __hash__ and __eq__ to MF_Base, and each subclass expects a hash_key method which will just be a tuple of its type and important attributes. Now sets actually work correctly for things like .get_all_variables().

MatchExpression can now work a lot better too as dictionary lookup is not instance-dependent anymore. Each template variable can now have a corresponding condition that an expression must pass in order to match with it. This is useful for things like the power rule only applying if the exponent is a number.... it has its limitations but it will be helpful for sure.