Actually working on the class I made this branch for lol.
An Address object will contain a list of addigits, which are either strings or integers. All the integers will always come after all the strings. Strings refer to the tree and which branch of the tree to go down, integers are simply the glyphs within that subexpression given by the strings.

This way we will be able to refer to, for example, the integer part of the number 123.456 as it is part of a larger expression. Maybe idk